requirements.txt
flask
tensorflow
numpy
Pillow
flask-cors
gunicorn
